Moving classes to their own files
