+++ Assets/Mirror/Tests/Editor/ClientSceneTests_ClearSpawners.cs
@@ -1,0 +1,66 @@
+﻿using System;
+using NUnit.Framework;
+
+namespace Mirror.Tests
+{
+    public class ClientSceneTests_ClearSpawners : ClientSceneTestsBase
+    {
+        [Test]
+        public void RemovesAllPrefabsFromDictionary()
+        {
+            prefabs.Add(Guid.NewGuid(), null);
+            prefabs.Add(Guid.NewGuid(), null);
+            prefabs.Add(Guid.NewGuid(), null);
+
+            ClientScene.ClearSpawners();
+
+            Assert.IsEmpty(prefabs);
+        }
+
+        [Test]
+        public void RemovesAllSpawnHandlersFromDictionary()
+        {
+            spawnHandlers.Add(Guid.NewGuid(), null);
+            spawnHandlers.Add(Guid.NewGuid(), null);
+            spawnHandlers.Add(Guid.NewGuid(), null);
+
+            ClientScene.ClearSpawners();
+
+            Assert.IsEmpty(spawnHandlers);
+        }
+
+        [Test]
+        public void RemovesAllUnspawnHandlersFromDictionary()
+        {
+            unspawnHandlers.Add(Guid.NewGuid(), null);
+            unspawnHandlers.Add(Guid.NewGuid(), null);
+            unspawnHandlers.Add(Guid.NewGuid(), null);
+
+            ClientScene.ClearSpawners();
+
+            Assert.IsEmpty(unspawnHandlers);
+        }
+
+        [Test]
+        public void ClearsAllDictionary()
+        {
+            prefabs.Add(Guid.NewGuid(), null);
+            prefabs.Add(Guid.NewGuid(), null);
+            prefabs.Add(Guid.NewGuid(), null);
+
+            spawnHandlers.Add(Guid.NewGuid(), null);
+            spawnHandlers.Add(Guid.NewGuid(), null);
+            spawnHandlers.Add(Guid.NewGuid(), null);
+
+            unspawnHandlers.Add(Guid.NewGuid(), null);
+            unspawnHandlers.Add(Guid.NewGuid(), null);
+            unspawnHandlers.Add(Guid.NewGuid(), null);
+
+            ClientScene.ClearSpawners();
+
+            Assert.IsEmpty(prefabs);
+            Assert.IsEmpty(spawnHandlers);
+            Assert.IsEmpty(unspawnHandlers);
+        }
+    }
+}
