+++ Assets/Mirror/Tests/Editor/ClientSceneTests_ClearSpawners.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ad2fe5633a9652045a5f7eb2643b6956
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
