+++ Assets/Mirror/Tests/Editor/ClientSceneTests_GetPrefab.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using NUnit.Framework;
+using UnityEngine;
+
+namespace Mirror.Tests
+{
+    public class ClientSceneTests_GetPrefab : ClientSceneTestsBase
+    {
+        [Test]
+        public void ReturnsFalseForEmptyGuid()
+        {
+            bool result = ClientScene.GetPrefab(new Guid(), out GameObject prefab);
+
+            Assert.IsFalse(result);
+            Assert.IsNull(prefab);
+        }
+
+        [Test]
+        public void ReturnsFalseForPrefabNotFound()
+        {
+            Guid guid = Guid.NewGuid();
+            bool result = ClientScene.GetPrefab(guid, out GameObject prefab);
+
+            Assert.IsFalse(result);
+            Assert.IsNull(prefab);
+        }
+
+        [Test]
+        public void ReturnsFalseForPrefabIsNull()
+        {
+            Guid guid = Guid.NewGuid();
+            prefabs.Add(guid, null);
+            bool result = ClientScene.GetPrefab(guid, out GameObject prefab);
+
+            Assert.IsFalse(result);
+            Assert.IsNull(prefab);
+        }
+
+        [Test]
+        public void ReturnsTrueWhenPrefabIsFound()
+        {
+            prefabs.Add(validPrefabGuid, validPrefab);
+            bool result = ClientScene.GetPrefab(validPrefabGuid, out GameObject prefab);
+
+            Assert.IsTrue(result);
+            Assert.NotNull(prefab);
+        }
+
+        [Test]
+        public void HasOutPrefabWithCorrectGuid()
+        {
+            prefabs.Add(validPrefabGuid, validPrefab);
+            ClientScene.GetPrefab(validPrefabGuid, out GameObject prefab);
+
+
+            Assert.NotNull(prefab);
+
+            NetworkIdentity networkID = prefab.GetComponent<NetworkIdentity>();
+            Assert.AreEqual(networkID.assetId, validPrefabGuid);
+        }
+    }
+}
