+++ Assets/Mirror/Tests/Editor/ClientSceneTests_RegisterPrefab.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bdfdcfafb85b3be418f2085e38663006
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
