+++ Assets/Mirror/Tests/Editor/ClientSceneTests_RegisterSpawnHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7e6577e8c2e64e41ae255edc61e91a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
