+++ Assets/Mirror/Tests/Editor/ClientSceneTests_UnregisterPrefab.cs
@@ -1,0 +1,54 @@
+﻿using NUnit.Framework;
+using UnityEngine;
+using UnityEngine.TestTools;
+
+namespace Mirror.Tests
+{
+    public class ClientSceneTests_UnregisterPrefab : ClientSceneTestsBase
+    {
+        [Test]
+        public void RemovesPrefabFromDictionary()
+        {
+            prefabs.Add(validPrefabGuid, validPrefab);
+
+            ClientScene.UnregisterPrefab(validPrefab);
+
+            Assert.IsFalse(prefabs.ContainsKey(validPrefabGuid));
+        }
+
+        [Test]
+        public void RemovesSpawnHandlerFromDictionary()
+        {
+            spawnHandlers.Add(validPrefabGuid, new SpawnHandlerDelegate(x => null));
+
+            ClientScene.UnregisterPrefab(validPrefab);
+
+            Assert.IsFalse(spawnHandlers.ContainsKey(validPrefabGuid));
+        }
+
+        [Test]
+        public void RemovesUnSpawnHandlerFromDictionary()
+        {
+            unspawnHandlers.Add(validPrefabGuid, new UnSpawnDelegate(x => { }));
+
+            ClientScene.UnregisterPrefab(validPrefab);
+
+            Assert.IsFalse(unspawnHandlers.ContainsKey(validPrefabGuid));
+        }
+
+        [Test]
+        public void ErrorWhenPrefabIsNull()
+        {
+            LogAssert.Expect(LogType.Error, "Could not unregister prefab because it was null");
+            ClientScene.UnregisterPrefab(null);
+        }
+
+        [Test]
+        public void ErrorWhenPrefabHasNoNetworkIdentity()
+        {
+            LogAssert.Expect(LogType.Error, $"Could not unregister '{invalidPrefab.name}' since it contains no NetworkIdentity component");
+            ClientScene.UnregisterPrefab(invalidPrefab);
+        }
+
+    }
+}
