+++ Assets/Mirror/Tests/Editor/ClientSceneTests_UnregisterPrefab.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa6fa692d80eaf8419e559c35034f016
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
