+++ Assets/Mirror/Tests/Editor/ClientSceneTests_UnregisterSpawnHandler.cs
@@ -1,0 +1,39 @@
+﻿using NUnit.Framework;
+
+namespace Mirror.Tests
+{
+    public class ClientSceneTests_UnregisterSpawnHandler : ClientSceneTestsBase
+    {
+        [Test]
+        public void RemovesSpawnHandlersFromDictionary()
+        {
+            spawnHandlers.Add(validPrefabGuid, new SpawnHandlerDelegate(x => null));
+
+            ClientScene.UnregisterSpawnHandler(validPrefabGuid);
+
+            Assert.IsFalse(unspawnHandlers.ContainsKey(validPrefabGuid));
+        }
+
+        [Test]
+        public void RemovesUnSpawnHandlersFromDictionary()
+        {
+            unspawnHandlers.Add(validPrefabGuid, new UnSpawnDelegate(x => { }));
+
+            ClientScene.UnregisterSpawnHandler(validPrefabGuid);
+
+            Assert.IsFalse(unspawnHandlers.ContainsKey(validPrefabGuid));
+        }
+
+        [Test]
+        public void DoesNotRemovePrefabDictionary()
+        {
+            prefabs.Add(validPrefabGuid, validPrefab);
+
+            ClientScene.UnregisterSpawnHandler(validPrefabGuid);
+
+            // Should not be removed
+            Assert.IsTrue(prefabs.ContainsKey(validPrefabGuid));
+        }
+
+    }
+}
