+++ Assets/Mirror/Tests/Editor/ClientSceneTests_UnregisterSpawnHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49ef76b8883ba3845942503683c9a9b3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
